+++ lib/screens/home.dart
@@ -1,6 +1,7 @@
 import 'package:albert_heijn_clone/components/appbar.dart';
 import 'package:albert_heijn_clone/components/home_ad.dart';
-import 'package:albert_heijn_clone/components/home_this_week.dart';
+import 'package:albert_heijn_clone/components/home_bonus_brochure_section.dart';
+import 'package:albert_heijn_clone/components/home_this_week_section.dart';
 import 'package:albert_heijn_clone/components/search_bar_button.dart';
 import 'package:albert_heijn_clone/localization/hardcoded.dart';
 import 'package:flutter/material.dart';
@@ -33,6 +34,7 @@ class HomeScreen extends StatelessWidget {
                 const SizedBox(
                   height: 28,
                 ),
+                const HomeBonusBrochureSection(),
                 Text(
                   'Welcome'.hardcoded,
                   style: const TextStyle(fontSize: 25),
